+++ .changeset/sweet-rocks-sing.md
@@ -1,0 +1,6 @@
+---
+"@gradio/wasm": minor
+"gradio": minor
+---
+
+feat:Remove Ruff and Uvicorn in Wasm env
